// src/app/interfaces/clientes.ts
export interface Clientes{
  id: string;
  numero_identificacion: string;
  primer_apellido: string;
  segundo_apellido: string;
  nombre: string;
  telefono1: string;
  telefono2: string;
  correo: string;
  sexo: string;
  fecha_nacimiento: string;
  fecha_creacion: string;
  fecha_modificacion: string;
  usu_creador: string;
  usu_modificador: string;
  estado: string;
  id_tid: string;
}